# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface MasterBase {
  id: number;
  code: string;
  name: string;
  phone?: string;
  address?: string;
  openingBalance?: number;
}

export type Customer = MasterBase;

export type Supplier = MasterBase;

export type Product = MasterBase & {
  unit: string;
  salePrice: number;
};

export type Material = MasterBase & {
  unit: string;
  cost: number;
};

export interface CustomerInput {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  name: string;
  phone?: string;
  address?: string;
  openingBalance?: number;
}

export type SupplierInput = CustomerInput;

export interface ProductInput {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  unit: string;
  /** @minimum 0 */
  salePrice: number;
}

export interface MaterialInput {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  unit: string;
  /** @minimum 0 */
  cost: number;
}

export interface TransactionItem {
  code: string;
  name: string;
  /** @exclusiveMinimum 0 */
  quantity: number;
  /** @minimum 0 */
  unitPrice: number;
  /** @minimum 0 */
  total: number;
}

export interface Sale {
  id: number;
  invoiceNo: string;
  date: string;
  customerName: string;
  total: number;
  paymentMethod: string;
  status?: string;
  items: TransactionItem[];
}

export interface SaleInput {
  invoiceNo: string;
  date: string;
  customerName: string;
  /** @minimum 0 */
  total: number;
  paymentMethod: string;
  /** @minItems 1 */
  items: TransactionItem[];
}

export interface Purchase {
  id: number;
  invoiceNo: string;
  date: string;
  supplierName: string;
  total: number;
  paymentMethod: string;
  status?: string;
  items: TransactionItem[];
}

export interface PurchaseInput {
  invoiceNo: string;
  date: string;
  supplierName: string;
  /** @minimum 0 */
  total: number;
  paymentMethod: string;
  /** @minItems 1 */
  items: TransactionItem[];
}

export type CashEntryType = typeof CashEntryType[keyof typeof CashEntryType];


export const CashEntryType = {
  قبض: 'قبض',
  صرف: 'صرف',
} as const;

export interface CashEntry {
  id: number;
  date: string;
  type: CashEntryType;
  amount: number;
  description: string;
  category?: string;
  partyType?: string;
  partyName?: string;
  analysis?: string;
  loading?: string;
  notes?: string;
  voucherNo?: string;
  accountCode?: string;
  accountName?: string;
  beneficiary?: string;
  analysisCode?: string;
  analysisName?: string;
  sourceType?: string;
  sourceId?: number;
  sourceNo?: string;
  balance: number;
}

export type CashEntryInputType = typeof CashEntryInputType[keyof typeof CashEntryInputType];


export const CashEntryInputType = {
  قبض: 'قبض',
  صرف: 'صرف',
} as const;

export interface CashEntryInput {
  date: string;
  type: CashEntryInputType;
  /** @exclusiveMinimum 0 */
  amount: number;
  /** @minLength 1 */
  description: string;
  category?: string;
  partyType?: string;
  partyName?: string;
  analysis?: string;
  loading?: string;
  notes?: string;
  voucherNo?: string;
  accountCode?: string;
  accountName?: string;
  beneficiary?: string;
  analysisCode?: string;
  analysisName?: string;
}

export interface InventoryRow {
  code: string;
  name: string;
  unit: string;
  purchasedQty: number;
  consumedQty: number;
  balanceQty: number;
  balanceValue: number;
  averageCost?: number;
}

export interface Partner {
  id: number;
  name: string;
  totalShare: number;
  paid: number;
  remaining: number;
  profitShare: number;
}

export interface Asset {
  id: number;
  name: string;
  purchaseValue: number;
  usefulLifeMonths: number;
  monthlyDepreciation: number;
  accumulatedDepreciation: number;
  bookValue: number;
}

export interface JournalLine {
  account: string;
  debit: number;
  credit: number;
}

export interface JournalEntry {
  id: number;
  date: string;
  description: string;
  source?: string;
  lines: JournalLine[];
}

export type DashboardRecentActivityItem = {
  date: string;
  label: string;
  amount: number;
  kind: string;
};

export interface Dashboard {
  salesTotal: number;
  purchasesTotal: number;
  cashBalance: number;
  inventoryValue: number;
  customerReceivables: number;
  supplierPayables: number;
  netActivity?: number;
  recentActivity: DashboardRecentActivityItem[];
}

export type ReconciliationOverallStatus = typeof ReconciliationOverallStatus[keyof typeof ReconciliationOverallStatus];


export const ReconciliationOverallStatus = {
  سليم: 'سليم',
  يحتاج_مراجعة: 'يحتاج مراجعة',
} as const;

export type ReconciliationChecksItemStatus = typeof ReconciliationChecksItemStatus[keyof typeof ReconciliationChecksItemStatus];


export const ReconciliationChecksItemStatus = {
  سليم: 'سليم',
  تحذير: 'تحذير',
  خطأ: 'خطأ',
} as const;

export type ReconciliationChecksItem = {
  label: string;
  status: ReconciliationChecksItemStatus;
  value: number;
  note: string;
};

export interface Reconciliation {
  overallStatus: ReconciliationOverallStatus;
  checks: ReconciliationChecksItem[];
}

export interface IncomeReport {
  revenue: number;
  directCosts: number;
  expenses: number;
  depreciation: number;
  netIncome: number;
}

/**
 * Record not found
 */
export type NotFoundResponse = Error;