# lib/api-zod/src/generated/types/cashEntry.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CashEntryType } from './cashEntryType';

export interface CashEntry {
  id: number;
  date: Date;
  type: CashEntryType;
  amount: number;
  description: string;
  category?: string;
  partyType?: string;
  partyName?: string;
  analysis?: string;
  loading?: string;
  notes?: string;
  voucherNo?: string;
  accountCode?: string;
  accountName?: string;
  beneficiary?: string;
  analysisCode?: string;
  analysisName?: string;
  balance: number;
}