# lib/api-zod/src/generated/types/cashEntryInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CashEntryInputType } from './cashEntryInputType';

export interface CashEntryInput {
  date: Date;
  type: CashEntryInputType;
  /** @exclusiveMinimum 0 */
  amount: number;
  /** @minLength 1 */
  description: string;
  category?: string;
  partyType?: string;
  partyName?: string;
  analysis?: string;
  loading?: string;
  notes?: string;
  voucherNo?: string;
  accountCode?: string;
  accountName?: string;
  beneficiary?: string;
  analysisCode?: string;
  analysisName?: string;
}