# lib/api-zod/src/generated/types/customer.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MasterBase } from './masterBase';

export type Customer = MasterBase;