# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardRecentActivityItem } from './dashboardRecentActivityItem';

export interface Dashboard {
  salesTotal: number;
  purchasesTotal: number;
  cashBalance: number;
  inventoryValue: number;
  customerReceivables: number;
  supplierPayables: number;
  netActivity?: number;
  recentActivity: DashboardRecentActivityItem[];
}