# lib/api-zod/src/generated/types/inventoryRow.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface InventoryRow {
  code: string;
  name: string;
  unit: string;
  purchasedQty: number;
  consumedQty: number;
  balanceQty: number;
  balanceValue: number;
  averageCost?: number;
}