# lib/api-zod/src/generated/types/partner.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Partner {
  id: number;
  name: string;
  totalShare: number;
  paid: number;
  remaining: number;
  profitShare: number;
}